+++ noodles-tabix/src/async/reader.rs
@@ -212,7 +212,9 @@ where
     let mut metadata = None;
 
     for _ in 0..n_bin {
-        let id = reader.read_u32_le().await?;
+        let id = reader.read_u32_le().await.and_then(|n| {
+            usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+        })?;
 
         if id == METADATA_ID {
             metadata = read_metadata(reader).await.map(Some)?;
@@ -289,7 +291,9 @@ where
 {
     use crate::index::reference_sequence::bin::METADATA_CHUNK_COUNT;
 
-    let n_chunk = reader.read_u32_le().await?;
+    let n_chunk = reader.read_u32_le().await.and_then(|n| {
+        usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+    })?;
 
     if n_chunk != METADATA_CHUNK_COUNT {
         return Err(io::Error::new(